divisione per inquinante e stazione
Right bois, we did it!

The code is really vrunchy and it works.
We have a dict and it contains the various pollution stations and it is accessible with the standard nomeclature i used in the code.

Nice work

Next step merging meteo and pollution then ML DF is done!

GG ALL